refactor Duration to avoid conflict with < in the new core (#478)
I'm working on the changes to introduce the new core design but I'm
having trouble with `<` now being a case class and the symbol
conflicting with `<` from the `Duration` extension methods. This PR
changes the encoding of `Duration` so the extension methods can be under
the companion object instead of the package to avoid the conflict.