prelude: add `evalNow` and `evalPartial` methods to the pending type (#516)
I'm using both methods to implement fibers based on `kyo-prelude`. 

The `evalNow` method inspects if the computation has no pending effects
to return its value if possible (via `Maybe`). In fibers, this method is
used to avoid forking computations that are already complete. I'm wasn't
sure about keeping this method public since it's low-level but it seems
useful to allow users and library authors to implement logic that avoids
overhead in case the computation is already complete.

The `evalPartial` is used from the `IOTask` effect handling to install
the interceptor and stop evaluation once the interceptor signals to stop
(preemption in fibers).