prelude: tracing fixes and improvements (#515)
I had thought that we wouldn't need to pass `Frame` explicit in methods
marked as `inline` but that's not the the case. Instead of using the
frame from the outer call, the compiler summons a new one. This PR
introduces the `Frame` implicit to inline methods, reduces the size of
the frame payload by inferring the short snippet from the expanded one,
and introduces a few other improvements (some unrelated to tracing)