prelude: add Safepoint.ensure (#517)
This PR introduces `Safepoint.ensure` to be used from `IO.ensure`. It
ensures the execution of a thunk at the end of the computation even if
the computation fails. In case an `Interceptor` is used to preempt
execution (like in `IOTask`), the interceptor can also collect the
thunks of the `ensure` calls to execute them later. It's a mechanism
[similar](https://github.com/getkyo/kyo/blob/14c1d921441bfd23279434559e7da9e5987e2c5f/kyo-core/shared/src/main/scala/kyo/ios.scala#L89)
to the current core.

This seems to be the last feature we need to port `kyo-core` to the new
design! \o/