prelude: Stream and Emit effects (#523)
This PR introduces a new `Emit` effect to be used as the underlying
effect of the new `Stream` implementation. The effect replaces `Sum`
since it has a very similar functionality.

The `Stream` implementation introduces two major changes:

1. The type now has only two parameters `Stream[S, V]` and the
underlying computation now produces an `Ack`. It's a simplification that
reduces the expressivity of the effect but I think usability is more
important. We can later add a separate implementation like `Flow[S, V,
A]` with a functionality like the previous Stream implementation if
needed. I placed the `S` parameter as the first one. It's similar to the
new `Fiber[E, A]` design.

2. The implementation now properly handles early termination
(https://github.com/getkyo/kyo/issues/368). I won't close the ticket for
now since I still need to add the `IO`-based operations as extension
methods when porting `kyo-core`.

I've also made a minor improvement to `Var.update` so it returns the
updated value.