prelude: Loop improvements (#535)
This PR introduces an optimization to specialize `Continue*` instances.
They're currently generic, which requires boxing of primitive types in
the `Continue` instance fields. After this change, the fields in the
final bytecode are specialized. Example from a decompiled loop:

```java
new Loop.Continue2<String, Object>(i1, i2){
    private final String i1$tailLocal1$1;
    private final int i2$tailLocal1$1;
```

There's still boxing when the value is recovered in the generic scope of
the `Loop` body but that's something the JIT can easily optimize.

I'm also using abbreviated generic names. It makes it easier for me to
read the code but I'd be ok to moving back to more descriptive names.
The `Loop.transform*` methods are renamed to `Loop.apply*` to improve
readability since `transform` doesn't seem to communicate much meaning
in this context.