prelude: avoid allocations for tracing (#536)
The new kernel allocates new traces on each `Boundary` use. This is
proving too expensive when porting fibers to the new design in
`kyo-prelude`. I've been avoiding caching state in the library since it
can become problematic depending on the characteristics of specific
workloads. For example, a cache might show great performance in a
benchmark but become a major bottleneck in real applications if it
doesn't have enough space and the hit ratio drops.

That said, this seems to be the first case where there's no alternative
to avoid allocations in such critical piece of code since fiber
execution is a key aspect of the library's performance. With this
change, traces are cached and reused as needed, which avoids the vast
majority of allocations. To reduce contention, each thread maintains a
local cache and replenishes from the global cache as necessary.