prelude: optimizations for IOTask (#548)
This seems to be the last optimizations needed to port `kyo-core` to the
new design. I had designed a generic `Boundary` mechanism that could be
used even by user-defined effects but the abstraction requires a few
allocations in the critical path of the the fiber execution (`IOTask`).
I'm kind of throwing the towel for now and providing more low-level APIs
designed specifically to ensure peak performance for fibers.