prelude: revert pending type encoding to an opaque type (#563)
I'm porting the other modules to the new design in `kyo-prelude` but
there's a performance regression that I can't see how to avoid with the
new encoding in the current version of the Scala 3 compiler:

- Given the lack of specialization, the generated bytecode contains a
significant amount of unexpected `<` boxing. Fortunately, the JIT is
able to avoid most of these allocations but profiling sessions show `<`
allocations in several benchmarks. It isn't a bottleneck in the
benchmarks I analyzed but introduces some overhead.

- The compiler doesn't allow implementing interfaces with `AnyVal`
generic type parameters. As a workaround, I was planning to use a
wrapper class like in `MonadLawsTest` but the approach is too expensive
in integrations like in `kyo-sttp`. The wrapper class essentially
introduces a new allocation to all methods in
[KyoSttpMonad](https://github.com/getkyo/kyo/blob/0279f30163759d834b1793cbba5bcf7d08f9f0c3/kyo-sttp/shared/src/main/scala/kyo/internal/KyoSttpMonad.scala#L9),
which are heavily used by sttp's transformations. Given how critical the
integrations with sttp and tapir are, this regression doesn't seem
ideal.

As described in https://github.com/getkyo/kyo/issues/531, the purpose of
the new `AnyVal` encoding was to allow arbitrary nesting of Kyo
computations. Although it's a nice usability improvement, the current
limitations of the Scala 3 compiler doesn't seem to make it worth it and
finalizing the migration to the new design seems a much higher priority.
I can create a ticket so we can try again later.

Related compiler tickets: https://github.com/scala/scala3/issues/11264
https://github.com/scala/scala3/issues/15532