fix readme (#577)
Following up on https://github.com/getkyo/kyo/pull/572, this PR fixes
the readme to incorporate the changes in the new core design. This is
only an initial pass fixing the code examples. I'm planning to follow up
with a more comprehensive documentation review but I think we can do
that after we have a release with the new design out.