make Chunk a Seq (#578)
This PR makes `Chunk` extend from `Seq` so we can avoid additional
allocations to transform between the types and improve usability. It's
not ideal since Scala's `Seq` methods aren't well-optimized but it seems
a good tradeoff.

The methods that used to take effectful functions like `map` and
`filter` were moved to `Kyo.*` and the methods from `Kyo.seq.*` are now
available directly in `Kyo.*`.