debug effect (#579)
This new effect uses the `Safepoint.Interceptor` mechanism introduced by
https://github.com/getkyo/kyo/pull/489 and the code snippets captured by
`Frame` to output debugging information. The `Debug.apply` method only
logs the result of the computation with its frame while `Debug.trace`
outputs all transformations of a computation with their values.

There's an important limitation: the `Safepoint.Interceptor` currently
isn't propagated to forked fibers so the debugging won't output the
trace of computations in forked fibers. I'll follow up on this
limitation but it's a bit tricky because we need to avoid increasing the
memory footprint of `IOTask`.

Example output of `Debug.trace`:


![image](https://github.com/user-attachments/assets/86f54716-891c-4613-a6a2-404b0ead27c4)