Refactor effect types (#586)
I'm working on documenting custom effects and the current naming using
`Effect` and `ContextEffect` is a bit confusing. The `Effect` name
doesn't communicate much meaning and I have to explain in the
documentation that it's an abstract arrow.

This PR refactors the effect types with an abstract `Effect` and two
sub-classes: `ArrowEffect` and `ContextEffect`.