tag: fix inference in generic scopes (#676)
Fixes #671.

This was a though one to debug. It seems something changed in the
compiler along the way and we didn't have tests for tags involving
generic parameters. The cause was `Expr.summon[Tag[tpe]]` in the macro
being interpreted by the compiler as `Expr.summon[String]` since it's in
a scope where `opaque type Tag` is known to be a `String`. The fix moves
the macro to another scope so it doesn't have visibility of the
underlying type of `Tag`.