prelude: allow union types in Abort.run (#708)
Fixes #698

My plan was to extend the functionality of `Tag` to allow matching tags
against runtime values in use cases like `Abort` but that's something
that will take time since we'd need a new implementation to encode union
and intersection types in the tag string and figure out a way to safely
compare tags to runtime values/classes.

This PR introduces a new `SafeClassTag` as a simpler alternative to
solve #698. It has a functionality similar to the standard library
`ClassTag` based on runtime classes but without quirks like falling back
to `Any` for unions/intersections and it also properly handles special
types like `Nothing` and `AnyVal`.

The macro doesn't allow tags for types that aren't known to be a class
type like in `Tag`, and, to ensure correctness, it's not possible to
obtain a `SafeClassTag` for parametrized types like `List[Int]`. This
means that it won't be possible to abort with parametrized types, which
I assume isn't very common. We should eventually be able to support
parametrized types with a new `Tag` implementation mentioned initially.