prelude: propagate safepoints in Loop (#720)
Kyo's `Safepoint` mechanism is responsible for providing stack safety,
tracing, and interceptors for features like preemption and the `Debug`
effect. The `Safepoint` is held in a `ThreadLocal` but, given that every
transformation in Kyo has to use it, always reading from the
`ThreadLocal` can introduce significant performance overhead. To
mitigate it, Kyo's kernel also propagates the safepoint via implicits
but the `Loop` implementation is missing it.