core: Async cleanups (#727)
The `Async.scala` file is getting too large and it's likely we'll have
to extend the provided APIs. This PR moves `Fiber` and `Promise` to a
separate file. I've also taken the opportunity to fix `fiber.block` to
use the provided `Clock` in the environment using the new unsafe APIs.