build: configure mima binary checks (#756)
Maintaining binary compatibility will be an important aspect once we
reach `1.0`. This PR introduces
[mima](https://github.com/lightbend-labs/mima) but configured to only
produce warnings for now. I'm planning to keep an eye on it to have a
better understanding of what will take to maintain binary compatibility
and identify patterns we could follow to make the process easier. We
should eventually also adopt
[tasty-mima](https://github.com/scalacenter/tasty-mima) but I couldn't
find a way to make it only warn of issues.