data: Text (#774)
This PR introduces `Text` in `kyo-data` to provide more efficient
handling of `String` operations. The main optimization is representing
substrings via a `Cut` wrapper instead of copying strings.

This is part of the work to introduce a new `Parse` effect in
`kyo-prelude` using effect handling for a more convenient backtracking
mechanism.