[core][sttp] fix timeout interruption (#797)
Easyracer is having trouble with Kyo not interrupting requests on
timeout. It's a bug similar to https://github.com/getkyo/kyo/pull/771.
`Async.timeout` is only propagating a completion, which isn't enough to
interrupt execution now that we split regular completions from
interruptions.

This PR changes `IOPromise.interrupt` and related methods to allow an
`Error[E]` so it's possible to interrupt the execution with a
`Result.Fail(timeoutError)` in `Async.timeout` instead of a `Panic`.