[prelude][core] isolated locals and reentrant meters (#775)
Kyo's `Meter`s are currently non-reentrant. That's not an ideal default
because it can easily lead to deadlocks. If a fiber tries to obtain the
same mutex multiple times for example, it won't ever resume. This PR is
the initial step to enable reentrancy by providing a new mechanism that
isolates locals so they aren't inherited by forked fibers.

The solution uses a new `ContextEffect.Isolated` marker trait to filter
entries from `Context` when a computation is being forked via a
`Boundary`. The `Local` implementation uses two effect types to
differentiate locals that are isolated or not.

With this new mechanism, I'm planning to change `Meter` to propagate an
isolated local indicating that the fiber is currently holding a permit
for a specific meter. Since locals are scoped, they will be cleared up
as soon as the computation is finished, right before the permit is
released. Forked fibers also won't see the local since it'll be
isolated. The goal is to allow users to customize if a meter should be
reentrant or not.