[prelude][core] Frame and Log optimizations (#800)
I noticed a major performance regression in `LoggingBench`. The main
cause is a bug introduced by https://github.com/getkyo/kyo/pull/783 that
makes the benchmark consider `Level.error` enabled when it isn't. I
fixed the bug but also worked on a few optimizations to improve
performance in case the log is enabled.

Benchmark
[results](https://jmh.morethan.io/?sources=https://gist.githubusercontent.com/fwbrasil/b7686518b8168f6fcba4ab2787d0ff41/raw/29e149f7bb688672cfdad844e308cde4df999e0e/jmh-result-baseline.json,https://gist.githubusercontent.com/fwbrasil/b7686518b8168f6fcba4ab2787d0ff41/raw/29e149f7bb688672cfdad844e308cde4df999e0e/jmh-result-candidate.json):


![image](https://github.com/user-attachments/assets/550bfe52-96dd-430f-a8b9-472f71d35b6e)