[prelude] introduce `Stream#rechunk` (#854)
PR 2 for #679 

Rechunk was a bit tricky to implement as the remaining 'buffer' would
get dropped unless the chunks perfectly aligned with the number of
emits. Because of this, it's necessary to Emit a final empty chunk to
ensure the buffer is flushed. This can sometimes drop the number of
elements in a given chunk if a user Emits an empty chunk, but that
shouldn't be a problem.

---------

Co-authored-by: Flavio Brasil <fwbrasil@users.noreply.github.com>