[prelude][core] effect isolates (#853)
Please see the `Isolate` scaladoc for more background. Ideally, we
should find a way to merge `Boundary` and `Isolate` but that's
challenging with the inference issues we've been having and it's
difficult to match the current performance of `Boundary` since it's
optimized by using more low-level kernel APIs. This new mechanism is
focused only on effects that maintain state but don't transform values.

Main benefits of this change:

- It allows users to fork computations with `Var`, `Emit`, `Memo`, and
`Check`.
- Users can leverage `isolate.run` directly to enforce the effect
handling for a computation
- Isolates provide transaction-like behavior via interaction with
`Abort`. This behavior will be necessary for an STM-based effect since
it'll need to rollback state to retry computations if necessary.

`Isolate` can't support effects like `Choice` since it transforms `A`
into `Seq[A]`.