[data] fix Duration in scala native (#867)
I'm having trouble to port support Scala Native in `kyo-core` because of
missing classes. The problem isn't shown in tests because zio tests are
disabled in Scala Native and the `zio-test-sbt` dependency adds the
necessary stubs.

This PR removes the dependency on `zio-test-sbt` and adds an explicit
dependency to `scala-java-time`.