[data] anchored and jittered schedules (#919)
I'm working on STM optimizations and want to use a jittered schedule to
reduce the impact of retry storms. Since we avoid side effects in
`kyo-data`, I'm passing the current `Instant` to `schedule.next` and
calculating the jitter deterministically via hashing. It isn't random
but, given that `Instant` has nanosecond precision, it seems reasonable.

We could pass a random number from `kyo-core` instead but we'd need to
eventually pass the instant anyway to implement more advanced schedules
like the new `Schedule.anchored` I've also added in this PR.