[prelude] move kernel to a new module (#934)
I'd like to do some experiments swapping Kyo's kernel with different
implementations mostly for educational purposes. It's currently
challenging with it being in `kyo-prelude`. With this change, it's
possible to just copy the kernel module and update the sbt dependencies
to test different implementations. The separation will also be helpful
to avoid the kernel depending on specific effects and to reduce
compilation times.

There are no behavior changes in this PR and a few tests had to be
adapted to not depend on prelude effects like `Env`.