[prelude][core][stm] fix issue with preemption in nested async computations (#927)
Fixes #925

The issue is more generic than the use with STM. Kyo's kernel uses a
safepoint mechanism to preempt execution on any step of a computation.
`kyo-core` extends this mechanism by setting the safepoint interceptor
to the current fiber to handle preemption. For example, if the fiber is
interrupted, the next safepoint will trigger preemption to stop the
computation and clear finalizers. The other possibility is the scheduler
itself triggering preemption due to the fiber taking more than the
configured time slice (10ms by default) so it can return to the
execution queue to provide fairness.

The cause of the bug is the interceptor propagating to nested async
computations. This PR fixes `Safepoint.eval`, which is called from the
pending type's `eval`, to clear the interceptor since the `eval`
operation itself can't be delayed and has to return a concrete value.
Even if the outer async computation is preempted, the nested one needs
to execute to completion.