[stm] use current transaction in initNow if available (#949)
`TRef.initNow` is essentially a footgun. If called within a transaction,
it'll make the transaction entry a retry loop for as long as the retry
`Schedule` allows. I spent some time debugging an issue caused by this
scenario because I added a transaction in an outer scope that had a
`Tref.initNow` in it.