[data] introduce Environment and improve KyoException (#1057)
<!--
PRs require an approval from any of the core contributors, other than
the PR author.

Include this header if applicable:
Fixes #issue1, #issue2, ...
-->

### Problem
<!--
Explain here the context, and why you're making this change. What is the
problem you're trying to solve?
-->
`KyoException` is too verbose in production environments. It's designed
to improve the development experience by showing errors with the `Frame`
information but, in a production environment, the overhead to log
failures can significantly contribute to performance degradation.

### Solution
<!--
Describe your solution. Focus on helping reviewers understand your
technical approach and implementation decisions.
-->
Introduce `Environment` that enables the verbose mode only if the user
explicitly enables the development mode or sbt entries are present in
the classpath, which should be the case for Kyo development environments
since we recommend both sbt and metals. The risk of false positives
seems low since the check is strict.

### Notes
<!--
Add any important additional information as bullet points, such as:
- Implementation details reviewers should know about
- Open questions and concerns
- Limitations
-->
I've also improved `KyoException` because I noticed a
`NullPointerException`.

---------

Co-authored-by: Ondra Pelech <ondra.pelech@gmail.com>