[data][prelude] Debug.trace fixes and improvements (#1065)
<!--
PRs require an approval from any of the core contributors, other than
the PR author.

Include this header if applicable:
Fixes #issue1, #issue2, ...
-->

### Problem
<!--
Explain here the context, and why you're making this change. What is the
problem you're trying to solve?
-->
The output of `Debug.trace` is confusing because it logs the
intermediate steps within a Kyo method as coming from the user-facing
frame. The rendering of `Frame` is also presenting the incorrect line
number.

### Solution
<!--
Describe your solution. Focus on helping reviewers understand your
technical approach and implementation decisions.
-->
Change the interceptor to discard intermediate values within a frame and
fix the line number rendering in `Frame`.