'Discarding' methods accept `Any` instead of `Unit` (#1070)
<!--
PRs require an approval from any of the core contributors, other than
the PR author.

Include this header if applicable:
Fixes #issue1, #issue2, ...
-->
Fixes #1067 

### Problem
<!--
Explain here the context, and why you're making this change. What is the
problem you're trying to solve?
-->
A lot of methods accept effects that are only used for their
side-effects and thus typically have a result type of `Unit`. This
requires users to frequently append `.unit` to the effects they pass to
these methods. This PR changes all such methods to accept effects in the
form `Any < S` to remove the need for this boilerplate.

### Solution
<!--
Describe your solution. Focus on helping reviewers understand your
technical approach and implementation decisions.
-->
Public discarding methods now accept `Any < S` effects and then convert
them to `Unit < S` using `.unit` prior to using them. Generally it is
hard to avoid calling `.unit` since these effects are typically passed
to methods that require a `Flat` result (`Any` is not guaranteed to be
flat.)

Here are the methods that have changed:
- kyo-combinators/Combinators
  - `ensuring` 
- kyo-combinators/Constructors
  - `Kyo.acquireRelease`
  - `Kyo.addFinalizer`
  - `Kyo.async`
- kyo-core/Clock
  - `Clock.repeat...`
- kyo-core/Fiber
  - `onComplete`
  - `onInterrupt`
- kyo-core/scheduler/IOPromise
  - `onComplete`
  - `onInterrupt`
- kyo-core/IO
  - `IO.ensure`
- kyo-core/Resource
  - `Resource.ensure`
  - `Resource.acquireRelease`
- kyo-core/Log
  - `Log.logWhen`
- kyo-kernal/Kyo
  - `Kyo.foreachDiscard`
- kyo-kernal/Loop
  - `repeat`
  - `forever` 
- kyo-kernal/Safepoint
  - `Safepoint.ensure` 
- kyo-prelude/Emit
  - `Emit.runForeach`
- kyo-prelude/Poll
  - `Poll.apply`
- kyo-prelude/Stream
  - `Stream.tap`
  - `Stream.tapChunk`
  - `Stream.foreach`
  - `Stream.foreachChunk`

### Notes
<!--
Add any important additional information as bullet points, such as:
- Implementation details reviewers should know about
- Open questions and concerns
- Limitations
-->